requirements.txt
# FastAPI and ASGI server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
asyncpg==0.30.0

# AWS S3
boto3==1.35.95
botocore==1.35.95

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic-settings==2.7.1

# Validation
pydantic==2.10.5
email-validator==2.2.0

# Utilities
python-dateutil==2.9.0.post0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1
faker==33.1.0

# Development
black==24.10.0
flake8==7.1.1
mypy==1.14.0
isort==5.13.2

# Logging and monitoring
python-json-logger==3.2.1
